\name{get.transliterations}
\alias{get.transliterations}
\title{Obtain table of transliterations}
\description{

}
\usage{
get.transliterations(language)
}
\arguments{
  \item{language}{

}
\author{
Richard Iannone
}
\examples{
# Get the rules for transliterating German into US-ASCII
get.transliterations("German")
}
