% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidecode.R
\name{unidecode}
\alias{unidecode}
\title{Transliterate unicode strings to ascii text}
\usage{
unidecode(data, language, changes_df = FALSE)
}
\arguments{
\item{data}{a vector of character data.}

\item{language}{the name of the source language for the text to be processed.}

\item{changes_df}{determines whether a text file with the changes made during text processing should be generated.}
}
\description{
Transliterate strings from containing Unicode text to ascii text.
}
