% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_transliterations}
\alias{get_transliterations}
\title{Obtain table of transliterations}
\usage{
get_transliterations(language)
}
\arguments{
\item{language}{the name of the source language for the text to be processed.}
}
\description{
This function returns a data frame with transliterations for a specific language, or, a large default set of transliterations that is language agnostic. The function is primarly called by the 'unidecode' function.
}

