\name{unidecode}
\alias{unidecode}
\title{Transliterate strings from containing Unicode text to ascii text}
\description{

}
\usage{
unidecode(data, language, changes_file = TRUE, interactive = FALSE)
}
\arguments{
  \item{data}{

}
  \item{language}{

}
  \item{changes_file}{

}
  \item{interactive}{

}
}
\author{
Richard Iannone
}
\examples{

}
